Don't put colons in error symlink names.
It caused in error when trying to use actions/upload-artifact@v4 in
GitHub Actions. It says certain characters are disallowed in order to
maintain compatibility with filesystems like NTFS.

Maybe we should even filter all our on-disk paths to remove this.
According to the error message, they are ":<>|*?\r\n